+++ frontend/src/main.tsx
@@ -1,17 +1,17 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import LoginPage from './pages/LoginPage.tsx'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
-import RegisterPage from './pages/RegisterPage.tsx'
 // import { AppProvider } from '@/components/context/app.context.tsx'
 // import ProtectedRoute from '@/components/auth/index.tsx'
 import { Toaster } from './components/ui/sonner.tsx'
 // import { SocketProvider } from './components/context/socket.context.tsx'
-import HomePage from './pages/HomePage.tsx'
+import HomePage from './pages/home.tsx'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import Layout from './layout.tsx'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
+import LoginPage from './pages/login.tsx'
+import RegisterPage from './pages/register.tsx'
 
 const router = createBrowserRouter([
   {
