+++ frontend/src/pages/home.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 
 const HomePage = () => {
   return (
